move astro integration to package
