+++ packages/start/package.json
@@ -38,6 +38,7 @@
       "import": "./src/index.tsx",
       "require": "./src/index.tsx"
     },
+    "./astro": "./src/start.js",
     "./client": "./src/client.ts",
     "./server": "./src/server.ts",
     "./package.json": "./package.json"
